make badges in single embeds also ignore touch events
this actually doesn't make a visible difference at this point which is
why i didn't change it before, but the behavior is still technically
weird without it. pressing on the badge does register a press on the
embedded image like expected, but it's not at the point of contact. this
is visible when an android ripple is added to the pressable enclosing
the image