tweak padding in treeview replies
Moved some paddings around to remove the gap that appears to the left
and right of treeview replies that was not covered by the ripple
previously.
Now the pressable itself extends from the rightmost thread border line
thingy all the way to the right edge of the screen.

I tried to make sure that the layout stays exactly the same as before,
which unfortunately required some annoying style logic...