Ask before navigating away from composer
This makes the 'Discard draft?' post show up if you click on a link
within the composer.

That fixes a bug where you write a reply to a post and then lose the
reply by clicking on the avatar of the person who you’re responding to.
You would not be prompted whether you want to discard your draft like
when clicking on **Cancel**. This bug also applies to hovering over the
avatar and clicking on one of the links there.