Add unfollow prompt to avoid accidental unfollow
This adds a prompt to ask before unfollowing accounts to avoid
accidentally unfollowing someone.

This is useful because unfollowing is a reversible but destructive
operation, so accidentally following someone can easily be reversed but
not vice versa because the account will be gone from your Following feed
and your account's follows.