feat(sign-up): Add tip about domain name handles
A point of confusion when registering for Bluesky is the use of "username". Many people signing up only know of Bluesky as a Twitter alternative and do not have context around the protocol. During the sign up process, a user may be unable to find an available username and be discouraged from signing up[1].

In the most technically correct world, the sign-up process would not use the term "username" and allow a user to choose between a `.bsky.social` domain or their own but that is much more complicated ux -- most users do not understand domain names, nor do they want to.

A good middle-ground is informing users about support for domain names (with the term "verify" to plant the seeds that domain names can be used for verification) while allowing them to choose a username.

I don't love the phrasing but "did you know?" is intended to _inform_ without instructing, i.e: a user reads it and learns something but doesn't think they need to take any specific action. So a user frustrated their chosen usernames are taken might think "ah! I know that domains can be used, I'll sign up with whatever as my username and then switch to my domain after".

[1] https://bsky.app/profile/luciferrari.tifosi.social/post/3lgl7s7n4ok2c