Update the ContentFilter setting for tenor gif search
No other media "upload" (it's... Not really an upload here I guess, but still) limits content at post time, it is instead restricted at read-time by the client/viewer.

Per the [tenor docs](https://tenor.com/gifapi/documentation#contentfilter) the current "high" setting restricts it to the equivalent of "G-rated" gifs which quite frankly seems absurdly restrictive considering the other content that is permitted. This change as currently written bumps it to pg-13. 

I personally would just turn it "off" ("R-rated") but thought that was less likely to be accepted.